LayoutTests/platform/mac/imported/w3c/web-platform-tests/mathml/relations/html5-tree/dynamic-childlist-002-expected.txt


FAIL Appending and removing children to maction assert_approx_equals: block position (child 0) expected -26 +/- 1 but got -70
PASS Appending and removing children to menclose
PASS Appending and removing children to merror
FAIL Appending and removing children to mfrac assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 6.84375
PASS Appending and removing children to mi
FAIL Appending and removing children to mmultiscripts assert_approx_equals: inline size expected 0 +/- 1 but got 42.5
PASS Appending and removing children to mn
PASS Appending and removing children to mo
FAIL Appending and removing children to mover assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 1.078125
PASS Appending and removing children to mpadded
PASS Appending and removing children to mphantom
FAIL Appending and removing children to mroot assert_approx_equals: inline position (child 0) expected 0 +/- 1 but got 1.03125
PASS Appending and removing children to mrow
PASS Appending and removing children to ms
PASS Appending and removing children to mspace
FAIL Appending and removing children to msqrt assert_approx_equals: inline size expected 12.703125 +/- 1 but got 16.640625
PASS Appending and removing children to mstyle
FAIL Appending and removing children to msub assert_approx_equals: block position (child 0) expected 0 +/- 1 but got -5.875
FAIL Appending and removing children to msubsup assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 2.375
FAIL Appending and removing children to msup assert_approx_equals: inline size expected 10.625 +/- 1 but got 0
PASS Appending and removing children to mtable
PASS Appending and removing children to mtext
FAIL Appending and removing children to munder assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 3
FAIL Appending and removing children to munderover assert_approx_equals: block position (child 0) expected 0 +/- 1 but got 6
PASS Appending and removing children to semantics
maction:
mfrac:
mmultiscripts:
mover:
mroot:
msqrt:
msub:
msubsup:
msup:
munder:
munderover:
